Move get_username functions to utils file
